+++ NEWS.md
@@ -20,6 +20,7 @@ private alleles are desired instead.
 
 ## Known Bugs
 * `calc_pairwise_ld()` can error if a specific comparison level has no SNPs.
+* `get.snpR.stats()` has issues returning an allele frequency matrix alongside other statistics. It returns *just* an allele frequency matrix perfectly fine.
 
 # snpR 1.2.8